# rhs-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the RHS Backend API documentation.",
        "title": "RHS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/hello": {
            "get": {
                "description": "Returns a greeting from the server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Show a hello message",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/countServices": {
            "get": {
                "description": "Returns the number of services stored at the path configured by SERVICE_PATH.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Count configured services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Service configuration could not be loaded"
                    }
                }
            }
        },
        "/getServices": {
            "get": {
                "description": "Returns the JSON document stored at the path configured by SERVICE_PATH.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Get service configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Service configuration could not be loaded"
                    }
                }
            }
        },
        "/getValues": {
            "get": {
                "description": "Returns the value stored under the supplied Redis key.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redis"
                ],
                "summary": "Get a Redis value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redis key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "406": {
                        "description": "Key is required"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Returns a pong response when the API is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check API health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/setValues": {
            "post": {
                "description": "Stores a value under the supplied Redis key without expiration.",
                "tags": [
                    "redis"
                ],
                "summary": "Set a Redis value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redis key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Value to store",
                        "name": "value",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Value stored"
                    },
                    "406": {
                        "description": "Key and value are required"
                    },
                    "500": {
                        "description": "Redis operation failed"
                    }
                }
            }
        }
    }
}